#12 Gestionar conferencias totalmente funcional.
Se han añadido los métodos PUT, GET (individual), GET (todos), DELETE (individual), DELETE (todos) a la gestión de conferencias.

+++ src/main/java/com/conferencias/tfg/repository/ConferenceRepository.java
@@ -4,8 +4,8 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 
 import com.conferencias.tfg.domain.Conference;
 
-public interface ConferenceRepository extends MongoRepository<Conference, Integer> {
+public interface ConferenceRepository extends MongoRepository<Conference, Long> {
 
-	Conference findById(int id);
+	Conference findById(long id);
 
 }
